Shuffle permission functions and types around a bit
