+++ next-frontend/src/providers/PermissionProvider.tsx
@@ -1,40 +1,18 @@
 "use client";
 
 import React, { createContext, useContext, useMemo } from "react";
-import { components } from "@/types/openapi";
 import { useQuery } from "@tanstack/react-query";
 import useAPI from "@/lib/wca/useAPI";
 import { useSession } from "next-auth/react";
-
-export interface PermissionFunctions {
-  canAccessPanel: (panel: string) => boolean;
-  canAdministerCompetition: (competition: string) => boolean;
-  canAttendCompetition: (competition: string) => boolean;
-  canOrganizeCompetitions: (competition: string) => boolean;
-  canEditDelegateReport: (competition: string) => boolean;
-  canViewDelegateAdminPage: (competition: string) => boolean;
-  canViewDelegateReport: (competition: string) => boolean;
-  canCreateGroup: (group: string) => boolean;
-  canEditGroup: (group: string) => boolean;
-  canReadGroupCurrent: (group: string) => boolean;
-  canReadGroupPast: (group: string) => boolean;
-  canRequestToEditProfile: (profile: string) => boolean;
-}
-
-type PermissionContext = {
-  permissions?: components["schemas"]["UserPermissions"];
-} & PermissionFunctions;
+import {
+  hydrateUserPermissions,
+  type PermissionContext,
+} from "@/lib/wca/permissions";
+import Loading from "@/components/ui/loading";
 
 const PermissionContext = createContext<PermissionContext | null>(null);
 
 export const usePermissions = () => useContext(PermissionContext);
-
-const allOrSpecificScope = (
-  item: string,
-  scope: components["schemas"]["CompetitionPermissions"],
-) => {
-  return scope === "*" || scope.includes(item);
-};
 
 export default function PermissionProvider({
   children,
@@ -43,104 +21,21 @@ export default function PermissionProvider({
 }) {
   const { data: session } = useSession();
   const api = useAPI();
-  const { data: request, isLoading } = useQuery({
+  const { data: rawPermissions, isLoading } = useQuery({
     enabled: Boolean(session),
     queryKey: ["permissions", session?.user?.id],
     queryFn: () => api.GET("/users/me/permissions"),
+    select: (data) => data.data,
   });
 
   const permissions: PermissionContext = useMemo(() => {
-    const rawPermissions = request?.data;
-
     return {
       permissions: rawPermissions,
-      canAccessPanel: (panel) =>
-        Boolean(
-          rawPermissions &&
-            rawPermissions.can_access_panels.scope.includes(panel),
-        ),
-      canAdministerCompetition: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_administer_competitions.scope,
-            ),
-        ),
-      canAttendCompetition: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_attend_competitions.scope,
-            ),
-        ),
-      canOrganizeCompetitions: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_organize_competitions.scope,
-            ),
-        ),
-      canEditDelegateReport: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_edit_delegate_report.scope,
-            ),
-        ),
-      canViewDelegateReport: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_view_delegate_report.scope,
-            ),
-        ),
-      canViewDelegateAdminPage: (competition) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              competition,
-              rawPermissions.can_view_delegate_admin_page.scope,
-            ),
-        ),
-      canCreateGroup: (group) =>
-        Boolean(
-          rawPermissions &&
-            rawPermissions.can_create_groups.scope.includes(group),
-        ),
-      canEditGroup: (group) =>
-        Boolean(
-          rawPermissions &&
-            rawPermissions.can_edit_groups.scope.includes(group),
-        ),
-      canReadGroupCurrent: (group) =>
-        Boolean(
-          rawPermissions &&
-            rawPermissions.can_read_groups_current.scope.includes(group),
-        ),
-      canReadGroupPast: (group) =>
-        Boolean(
-          rawPermissions &&
-            rawPermissions.can_read_groups_past.scope.includes(group),
-        ),
-      canRequestToEditProfile: (profile) =>
-        Boolean(
-          rawPermissions &&
-            allOrSpecificScope(
-              profile,
-              rawPermissions.can_request_to_edit_others_profile.scope,
-            ),
-        ),
+      ...hydrateUserPermissions(rawPermissions),
     };
-  }, [request]);
+  }, [rawPermissions]);
 
-  if (isLoading) {
-    return <p>Loading...</p>;
-  }
+  if (isLoading) return <Loading />;
 
   return (
     <PermissionContext.Provider value={permissions}>
